requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.3
python-dotenv==1.0.0
pyjwt==2.7.0
bcrypt==4.0.1
openai==0.27.8
python-multipart==0.0.6